Windows: Carry over error code from synchronous transfer
GetOverlappedResult() correctly returns the error code from asynchronous
transfer completion, but when we try to get the synchronous transfer
error code through, it gets lost somewhere. It is unclear to me what
GetOverlappedResult() returns in all cases and where or whether the
overlapped Internal member is looked at.

This experimental commit is maybe more of a workaround if it works, but
will in any case provide more debug information.

Signed-off-by: Tormod Volden <debian.tormod@gmail.com>